move getImportAlias to utils
