+++ packages/components/codemods/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./transformSource"
+export * from "./getImportAlias"
